feat: add anchor for stakeholders widget
